fix: likeCount 타입 long으로 변경

+++ src/main/java/redlightBack/Post/Dto/PostLikeResponse.java
@@ -2,5 +2,5 @@ package redlightBack.Post.Dto;
 
 public record PostLikeResponse(Long postId,
                                boolean liked,
-                               int likeCount) {
+                               long likeCount) {
 }
